traffic_env/server/requirements.txt
# ── Core Framework ──
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ── OpenEnv ──
openenv-core>=0.1.0

# ── Inference (OpenAI client for baseline script) ──
openai>=1.12.0

# ── HTTP client (for inference.py hitting the FastAPI server) ──
httpx>=0.25.0
requests>=2.31.0

# ── Utilities ──
numpy>=1.24.0
